+++ .github/workflows/build.yml
@@ -29,4 +29,4 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: gradle/wrapper-validation-action@v1
-      - run: ./gradlew :build-logic:build
+      - run: ./gradlew :fabric-filament:build
