--- build-logic/src/test/resources/projects/javadocLint/build.gradle
@@ -1,7 +1,0 @@
-plugins {
-	id 'build-logic'
-}
-
-javadocLint {
-	mappingDirectory = file('mappings')
-}
