--- build-logic/src/test/resources/projects/unpickDef/build.gradle
@@ -1,13 +1,0 @@
-plugins {
-	id 'build-logic'
-}
-
-combineUnpickDefinitions {
-	input = file('unpick-definitions')
-	output = file('combined_definitions.unpick')
-}
-
-remapUnpickDefinitionsIntermediary {
-	output = file('intermediary_definitions.unpick')
-	mappings = file('yarn-mappings-v2.tiny')
-}
