+++ build.gradle
@@ -24,7 +24,7 @@ plugins {
 	id 'de.undercouch.download' version '4.1.2'
 	id 'maven-publish'
 	id "com.diffplug.spotless" version "6.4.2"
-	id 'build-logic'
+	id 'net.fabricmc.filament'
 }
 
 def minecraft_version = "1.19.2"
