+++ filament/src/test/resources/projects/javadocLint/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+	id 'net.fabricmc.filament'
+}
+
+javadocLint {
+	mappingDirectory = file('mappings')
+}
