+++ filament/src/test/resources/projects/sharedData/yarn-mappings-v2.tiny
