+++ filament/src/test/resources/projects/unpickDef/build.gradle
@@ -1,0 +1,13 @@
+plugins {
+	id 'net.fabricmc.filament'
+}
+
+combineUnpickDefinitions {
+	input = file('unpick-definitions')
+	output = file('combined_definitions.unpick')
+}
+
+remapUnpickDefinitionsIntermediary {
+	output = file('intermediary_definitions.unpick')
+	mappings = file('yarn-mappings-v2.tiny')
+}
