+++ settings.gradle
@@ -16,4 +16,4 @@ if (!JavaVersion.current().isCompatibleWith(JavaVersion.VERSION_17)) {
 
 rootProject.name = "yarn"
 
-includeBuild 'build-logic'
+includeBuild 'filament'
